+++ README.dev
@@ -12,16 +12,16 @@ format
 2) Add your new potential to
 potential_src/potential_c_ext/galpy_potentials.h
 
-3) Edit the code under orbit_src/orbit_c_ext/integratePlanarOrbit.c to
+3) Edit the code under orbit/orbit_c_ext/integratePlanarOrbit.c to
 set up your new potential (in the 'parse_leapFuncArgs' function)
 
-4) Edit the code in orbit_src/integratePlanarOrbit.py to set up your
+4) Edit the code in orbit/integratePlanarOrbit.py to set up your
 new potential
 
-5) Edit the code under orbit_src/orbit_c_ext/integrateFullOrbit.c to
+5) Edit the code under orbit/orbit_c_ext/integrateFullOrbit.c to
 set up your new potential (in the 'parse_leapFuncArgs_Full' function)
 
-6) Edit the code in orbit_src/integrateFullOrbit.py to set up your
+6) Edit the code in orbit/integrateFullOrbit.py to set up your
 new potential
 
 7) Edit the code in actionAngle/actionAngle_c_ext/actionAngle.c to
